Replaced `cs1302.hw03` with `cs1302.hw`.
The assignment is no longer tied to a particular number. This change will give
us more flexibility regarding when we assign it in subsequent semesters.

For prosperity, here is the command that was used to replace the strings:

```sh
grep -rl "cs1302\.hw03" | xargs sed -i 's|cs1302.hw03|cs1302.hw|g'
```